Users/rle/user to db -> Database Manipulation + Unit testing (#14)
* Add user OK, No hash password, no JPA Test

* users can be added to the db

* Removing DAO objects + Adding lombok for user + Setting some long everywhere + clean code

* properties reinitialized

+++ src/main/java/com/esgi/group5/jeeproject/controllers/OpinionController.java
@@ -29,7 +29,7 @@ public class OpinionController {
 
     @PostMapping
     public ResponseEntity<?> post(HttpServletRequest request, @RequestBody @Valid Opinion opinion){
-        int id = service.add(opinion);
+        long id = service.add(opinion);
 
         return ResponseEntity
                 .status(HttpStatus.CREATED)
