+++ src/main/java/com/esgi/group5/jeeproject/repositories/contracts/IUserRepository.java
@@ -4,8 +4,8 @@ import com.esgi.group5.jeeproject.models.User;
 
 import java.util.List;
 
-public interface IUserRepository {
-    int add(User user);
+public interface IUserRepository{
+    long add(User user);
     List<User> get();
-    User get(int id);
+    User get(long id);
 }
