+++ src/main/java/com/esgi/group5/jeeproject/services/contracts/ITradeService.java
@@ -5,7 +5,7 @@ import com.esgi.group5.jeeproject.models.Trade;
 import java.util.List;
 
 public interface ITradeService {
-    int add(Trade trade);
+    long add(Trade trade);
     List<Trade> get();
-    Trade get(int tradeId);
+    Trade get(long tradeId);
 }
